secure-using-steganograph/requirements.txt
bcrypt==3.1.7
cffi==1.13.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
Pillow==6.2.0
pycparser==2.19
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
opencv-python-headless
psycopg2==2.8.3

